finishing rest app

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require("path");
 const { validationResult } = require("express-validator");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -56,7 +57,7 @@ exports.createPost = (req, res, next) => {
 
   console.log(req.file);
   console.log(req.body);
-
+  let creator;
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
@@ -65,7 +66,7 @@ exports.createPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Vedran" }
+    creator: req.userId
   });
 
   if (imageUrl !== post.imageUrl) {
@@ -75,11 +76,19 @@ exports.createPost = (req, res, next) => {
   post
     .save()
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
 
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: "Post created succesfuly",
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -146,6 +155,16 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error(`Not authorized`);
+        error.statusCode = 403;
+        throw error;
+      }
+
+      if (imageUrl !== post.imageUrl) {
+        clearImage(post.imageUrl);
+      }
+
       post.title = title;
       post.content = content;
       post.imageUrl = imageUrl;
@@ -176,13 +195,24 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error(`Not authorized`);
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
 
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
-
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.post.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Post Deleted" });
     })
     .catch(err => {
